Blackerz-API/requirements.txt
aniso8601==4.1.0
Click==7.0
Flask==1.0.2
Flask-RESTful==0.3.7
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
pytz==2018.9
six==1.12.0
Werkzeug==0.14.1
requests==2.25.1
python-firebase==1.2.0
